[DTSSTCI-1293] Add check for draft order document in all affected WA tests
