+++ src/functionalTest/java/uk/gov/hmcts/sptribs/testutil/FunctionalTestSuite.java
@@ -10,6 +10,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.ClassPathResource;
+import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
 import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
@@ -20,6 +21,7 @@ import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
 import uk.gov.hmcts.reform.ccd.document.am.model.Classification;
 import uk.gov.hmcts.reform.ccd.document.am.model.DocumentUploadRequest;
 import uk.gov.hmcts.reform.ccd.document.am.util.InMemoryMultipartFile;
+import uk.gov.hmcts.sptribs.cdam.model.Document;
 import uk.gov.hmcts.sptribs.cdam.model.UploadResponse;
 import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
 import uk.gov.hmcts.sptribs.ciccase.model.DssCaseData;
@@ -37,6 +39,7 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
+import java.util.UUID;
 
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
@@ -49,6 +52,9 @@ public abstract class FunctionalTestSuite {
 
     private static final LocalDateTime LOCAL_DATE_TIME = LocalDateTime.of(2021, 4, 28, 1, 0);
     private static final Logger log = LoggerFactory.getLogger(FunctionalTestSuite.class);
+
+    private static final ClassPathResource DRAFT_ORDER_FILE =
+            new ClassPathResource("files/DRAFT :Order--[Subject Name]--26-04-2024 10:09:12.pdf");
 
     @Value("${test-url}")
     protected String testUrl;
@@ -319,7 +325,31 @@ public abstract class FunctionalTestSuite {
             .build();
     }
 
-    protected UploadResponse uploadTestDocument(ClassPathResource resource) {
+    protected void checkAndUpdateDraftOrderDocument(Map<String, Object> caseData) {
+        UploadResponse uploadResponse = uploadTestDocumentIfMissing("5d76ff31-8547-4702-b2c8-34c43a53d220", DRAFT_ORDER_FILE);
+
+        if (uploadResponse != null) {
+            log.info("Document uploaded: {}", uploadResponse.getDocuments().getFirst());
+            updateOrderTemplate(uploadResponse.getDocuments().getFirst(), caseData);
+        }
+    }
+
+    protected UploadResponse uploadTestDocumentIfMissing(String documentId, ClassPathResource resource) {
+        if (!checkDocumentExists(documentId)) {
+            return uploadTestDocument(resource);
+        }
+        return null;
+    }
+
+    private boolean checkDocumentExists(String documentId) {
+        final String serviceToken = serviceAuthenticationGenerator.generate();
+        final String userToken = idamTokenGenerator.generateIdamTokenForSystemUser();
+        ResponseEntity<Document> documentResponse = caseDocumentClientApi.getDocument(userToken, serviceToken, UUID.fromString(documentId));
+
+        return documentResponse.getStatusCode().is2xxSuccessful();
+    }
+
+    private UploadResponse uploadTestDocument(ClassPathResource resource) {
         cdamUrlDebugger.logUrls();
         log.debug("uploading FT test document");
         final List<AppsConfig.AppsDetails> appDetails = appsConfig.getApps();
@@ -351,5 +381,32 @@ public abstract class FunctionalTestSuite {
             }
         }
         return null;
+    }
+
+    @SuppressWarnings("unchecked")
+    private void updateOrderTemplate(Document document, Map<String, Object> caseData) {
+        Map<String, Object> orderTemplateIssued = (Map<String, Object>) caseData.get("cicCaseOrderTemplateIssued");
+        if (orderTemplateIssued != null) {
+            orderTemplateIssued.put("document_url", document.links.self.href);
+            orderTemplateIssued.put("document_filename", document.originalDocumentName);
+            orderTemplateIssued.put("document_binary_url", document.links.binary.href);
+        }
+        caseData.put("cicCaseOrderTemplateIssued", orderTemplateIssued);
+
+        List<Map<String, Object>> draftOrderList = (List<Map<String, Object>>) caseData.get("cicCaseDraftOrderCICList");
+        if (draftOrderList != null && !draftOrderList.isEmpty()) {
+            Map<String, Object> firstItem = draftOrderList.getFirst();
+            Map<String, Object> value = (Map<String, Object>) firstItem.get("value");
+            if (value != null) {
+                Map<String, Object> templateGeneratedDocument = (Map<String, Object>) value.get("templateGeneratedDocument");
+                if (templateGeneratedDocument != null) {
+                    templateGeneratedDocument.put("document_url", document.links.self.href);
+                    templateGeneratedDocument.put("document_filename", document.originalDocumentName);
+                    templateGeneratedDocument.put("document_binary_url", document.links.binary.href);
+                }
+                value.put("templateGeneratedDocument", templateGeneratedDocument);
+            }
+        }
+        caseData.put("cicCaseDraftOrderCICList", draftOrderList);
     }
 }
