+++ src/functionalTest/java/uk/gov/hmcts/sptribs/wa/WAProcessDirectionsReListedCaseFT.java
@@ -68,8 +68,8 @@ public class WAProcessDirectionsReListedCaseFT extends FunctionalTestSuite {
             CASEWORKER_RECORD_LISTING, ST_CIC_JURISDICTION, ST_CIC_CASE_TYPE, newCaseId, caseData);
 
         hearingCaseData.put("cicCaseReferralTypeForWA", "Listed case");
-
         hearingCaseData.putAll(caseData(CASEWORKER_CREATE_DRAFT_ORDER_DATA));
+        checkAndUpdateDraftOrderDocument(caseData);
         ccdCaseCreator.createInitialStartEventAndSubmit(
             CASEWORKER_CREATE_DRAFT_ORDER, ST_CIC_JURISDICTION, ST_CIC_CASE_TYPE, newCaseId, hearingCaseData);
