+++ src/functionalTest/java/uk/gov/hmcts/sptribs/wa/WAProcessDirectionsReListedCaseWithin5DaysFT.java
@@ -69,6 +69,7 @@ public class WAProcessDirectionsReListedCaseWithin5DaysFT extends FunctionalTest
 
         hearingCaseData.put("cicCaseReferralTypeForWA", "Listed case (within 5 days)");
         hearingCaseData.putAll(caseData(CASEWORKER_CREATE_DRAFT_ORDER_DATA));
+        checkAndUpdateDraftOrderDocument(caseData);
         ccdCaseCreator.createInitialStartEventAndSubmit(
             CASEWORKER_CREATE_DRAFT_ORDER, ST_CIC_JURISDICTION, ST_CIC_CASE_TYPE, newCaseId, hearingCaseData);
