+++ src/functionalTest/java/uk/gov/hmcts/sptribs/wa/WAProcessDirectionsReturnedFT.java
@@ -62,6 +62,7 @@ public class WAProcessDirectionsReturnedFT extends FunctionalTestSuite {
 
         caseData.put("cicCaseReferralTypeForWA", "New case");
         caseData.putAll(caseData(CASEWORKER_CREATE_DRAFT_ORDER_DATA));
+        checkAndUpdateDraftOrderDocument(caseData);
         ccdCaseCreator.createInitialStartEventAndSubmit(
             CASEWORKER_CREATE_DRAFT_ORDER, ST_CIC_JURISDICTION, ST_CIC_CASE_TYPE, newCaseId, caseData);
