+++ src/functionalTest/java/uk/gov/hmcts/sptribs/wa/WAProcessListingDirectionsFT.java
@@ -62,6 +62,7 @@ public class WAProcessListingDirectionsFT extends FunctionalTestSuite {
 
         caseData.put("cicCaseReferralTypeForWA", "Listing directions");
         caseData.putAll(caseData(CASEWORKER_CREATE_DRAFT_ORDER_DATA));
+        checkAndUpdateDraftOrderDocument(caseData);
         ccdCaseCreator.createInitialStartEventAndSubmit(
             CASEWORKER_CREATE_DRAFT_ORDER, ST_CIC_JURISDICTION, ST_CIC_CASE_TYPE, newCaseId, caseData);
