+++ src/functionalTest/java/uk/gov/hmcts/sptribs/wa/WAProcessPostponementDirectionsFT.java
@@ -69,6 +69,7 @@ public class WAProcessPostponementDirectionsFT extends FunctionalTestSuite {
 
         hearingCaseData.put("cicCaseReferralTypeForWA", "Postponement request");
         hearingCaseData.putAll(caseData(CASEWORKER_CREATE_DRAFT_ORDER_DATA));
+        checkAndUpdateDraftOrderDocument(caseData);
         ccdCaseCreator.createInitialStartEventAndSubmit(
             CASEWORKER_CREATE_DRAFT_ORDER, ST_CIC_JURISDICTION, ST_CIC_CASE_TYPE, newCaseId, hearingCaseData);
