+++ src/functionalTest/java/uk/gov/hmcts/sptribs/wa/WAProcessReinstatementDecisionNoticeFT.java
@@ -65,6 +65,7 @@ public class WAProcessReinstatementDecisionNoticeFT extends FunctionalTestSuite 
 
         caseData.put("cicCaseReferralTypeForWA", "Reinstatement request");
         caseData.putAll(caseData(CASEWORKER_CREATE_DRAFT_ORDER_DATA));
+        checkAndUpdateDraftOrderDocument(caseData);
         ccdCaseCreator.createInitialStartEventAndSubmit(
             CASEWORKER_CREATE_DRAFT_ORDER, ST_CIC_JURISDICTION, ST_CIC_CASE_TYPE, newCaseId, caseData);
