+++ src/functionalTest/java/uk/gov/hmcts/sptribs/wa/WAProcessRule27DecisionFT.java
@@ -62,6 +62,7 @@ public class WAProcessRule27DecisionFT extends FunctionalTestSuite {
 
         caseData.put("cicCaseReferralTypeForWA", "Rule 27 request");
         caseData.putAll(caseData(CASEWORKER_CREATE_DRAFT_ORDER_DATA));
+        checkAndUpdateDraftOrderDocument(caseData);
         ccdCaseCreator.createInitialStartEventAndSubmit(
             CASEWORKER_CREATE_DRAFT_ORDER, ST_CIC_JURISDICTION, ST_CIC_CASE_TYPE, newCaseId, caseData);
